deployment/requirements.txt
# Food Spoilage Detection API — runtime dependencies

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Data validation
pydantic>=2.0.0

# ML / vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
scikit-learn>=1.3.0

# Image handling
pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
